+++ addon/macros/to-lat-lng.js
@@ -5,7 +5,7 @@ export default function(latKey = 'lat', lngKey = 'lng') {
   return computed(latKey, lngKey, {
     get() {
       let [lat, lng] = [this.get(latKey), this.get(lngKey)];
-      return this.L.latLng(lat, lng);
+      return this.get('L').latLng(lat, lng);
     },
     set(key, value) {
       this.setProperties({